cerbero/build-tools: Transfer vs_install_* settings to build tools
Otherwise custom vs install locations passed via `-c` to cerbero
aren't set on the build tools config. In practice, this leads to .lib
files not being generated. In theory, this also causes build tools
recipes to not be buildable with Visual Studio, which might be
a problem in the future.