Remove unnecessary `.` at the end of cmake command
`.` at the end of cmake command is not necessary because the build and
work directory are explicitly specified with -H and -B options. In
cmake version 3.10.2 and older cmake 3 version  dot at the end leads
to ovewriting the value of -H option.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/484>