recipe: install licensing with config_src_dir as a base folder
In order to separate build folder from src folder, the license
will be taken from config_src_dir.

In the case of librtmp, the license is not in the config_src_dir where
the CMakeLists.txt is located but one level up.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/707>