build: fix the cmake template
In order to support a different build folder from
a src folder where srcdir is not the root of the package
folder, we need to to differentiate these 3 variables.

-H will use make_dir(where cmakelist.txt is)
-B will use build_dir where the build will be performed.
make_dir after build_dir after configure because the Makefile
is actually in the build_dir.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/488>