gst-plugins-bad-1.0.recipe: Disable wasapi2 plugin for now
The libraries are getting auto-detected and enabled, but the build
fails because we set the target Windows version as Windows 7.

We can only enable it for UWP till we figure out how we want to build
it and ship it w.r.t build-time and runtime backwards compat.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/502>