openssl.recipe: Ensure that Perl uses forward slashes
Someone reported on IRC that openssl was failing to build because it
was complaining about Perl using backward slashes instead of forward
slashes. Turned out to be because someone had overriden TERM to be
`dumb` instead of the default `cygwin`.

Set those vars ourselves to prevent this from happening.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/510>