custom.py: Make enable_plugin work like disable_plugin
If the `option` argument is not specified, `disable_plugin()` sets it
to `plugin`, but `enable_plugin()` sets it to `variant`.

This made sense in the very beginning because of some edge cases such
as v4l2 and pulse, but there's just as many cases where the option is
the same as the plugin name and that is true for the majority of
plugins, so make it consistent.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/511>