Add a recipe for using ANGLE with UWP
This is the binary used by the Servo folks, and is the best-maintained
ANGLE-for-UWP fork out there.

Instead of building it, we currently just download the nuget, extract
it, and put it into the prefix. There's just one nuget for ARM64 and
x86_64, but there's separate nugets for debug-crt and release-crt.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/538>