packages: Fix fetch-package + package --offline on Linux
On Linux, we will always add bash-completion to the list of recipes,
but because fetch-package maintains a separate list based on the list
of recipes in the package files, we will not fetch it. This is a known
bug. See: https://gitlab.freedesktop.org/gstreamer/cerbero/-/issues/125

Workaround it by adding it to -core.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/561>