gst-plugins-bad: Enable nvcodec by default
We do not need to compile against any Nvidia SDKs now, so we can build
this by default and ship it out of the box in our binary releases.

It's still disabled on UWP, iOS, and Android since it's not usable on
those.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/587>