Split bootstrap commands into system/toolchains/build-tools
This allows us to run the --system command in the docker file when
building the image, and --toolchains and --build-tools when running
a CI job.

This will simplify the job scripts, and will also allow us to test
changes to the toolchain code in cerbero without rebuilding the image.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/611>