bootstrap: Reclassify bootstrappers as system or toolchain
Some bootstrappers are responsible for bootstrapping the system
(usually the 'build distro' bootstrappers), and others are responsible
for bootstrapping the toolchain that will be used for building
(usually the 'target distro' bootstrappers).

We need to differentiate this so we can split the Windows bootstrapper
into separate classes for the toolchain (MinGW + other deps, on
Native-Windows and Cross-Windows) and MSYS (only on Native-Windows).

This also fixes downloading and extracting of toolchains in bootstrap
which broke in 49ab89744de42e6f3dbb13e1090b75371b5a8000

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/616>