windows.config: Use MinGW Perl instead of MSYS Perl
It seems like MSYS Perl sometimes hangs on the Windows CI runners and
causes the cerbero build jobs to hang. Maybe the MinGW Perl won't hang
because it's not ancient.

Some recipes such as autoconf and automake fail to build when using
the newer Perl, so continue to use the MSYS perl for those.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/616>