cerbero: Don't try to git commit if nothing to commit
git commit will fail if there's nothing to commit. This can happen if
all the commands in the extract step completed successfully, but the
step was marked as failed or was cancelled. In that case, the step
will be tried again and since all commands will be idempotent, there
won't be anything to commit.

Check whether we need to commit before trying to commit.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/784>