python: add support for multiarch distributions
In multiarch distributions, python module can be installed
in the purelib path or the platlib path.

For example in Fedora x86_64
  * lib/python-3.8/site-packages
  * lib64/python-3.8/site-packages

This patch tries to correctly support these scenarios by adding as python
prefixes both the purelib and the platlib paths.

It also provides a better mechanism to list python files in recipes
introducing the replacements `pydir` and `pyplatdir`

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/639>