bootstrap: Fix cross-windows bootstrap with --system-only
When we're bootstrapping on Linux for cross-windows, we don't want
"Windows" as a system bootstrapper. So we shouldn't be looping over
all the system bootstrappers. We should just pick the one that matches
the build distro. Same for the toolchain, because there can only be
at most one that matches the target distro.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/647>