Fix build of autotools projects needing macros from build-tools
In Windows, the bootstrap used to build all the autotools and now it's
using the ones from the msys installation. During the autogen phase,
GTK+ requires the pkg-config macros that are installed in the
build-tools prefix. Since we use the msys autotools, the macros in
/share/aclocal are no longer in the default search paths. This patch
makes sure that the m4 macros installed by a build tool can be used.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/671>