recipes/devtools/ges: fix universal packaging
Directory lists only work when they are the only package wanting to
contain files there.  If multiple recipes specify a directory in the
file list, the second is ignored which means that files are not copied
from the arch-speific directory to the final universal prefix.

Fixes package with macos universal

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/701>