ci: Fix deps build on pushes to main
The YAML was configured to fetch a cache and use it whenever a cerbero
pipeline is triggered via the GitLab pipeline trigger API, which
happens in the "cerbero trigger" job in gstreamer/gstreamer, which
runs gstreamer.git/ci/gitlab/trigger_cerbero_pipeline.py

But the job definition in Cerbero CI couldn't differentiate between
a pipeline triggered that way and a pipeline triggered as post-merge
CI to generate a Cerbero cache. The variable used to check that was
CI_PIPELINE_SOURCE which is 'pipeline' in both those cases:

https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/18446678 (post-merge CI)
https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/14644527 (triggered)

It is 'merge_request_event' when the pipeline is triggered by
a Cerbero merge request:

https://gitlab.freedesktop.org/ystreet/cerbero/-/jobs/18433849

The only way to differentiate between the two is to add a custom
trigger variable when triggering the pipeline from the gstreamer CI
job: CI_GSTREAMER_TRIGGERED

We set it to "false" for now, which will cause all pipelines to to
build from scratch. A corresponding MR on gstreamer/gstreamer will set
it to "true" which will use the cache.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/800>