osx/universal: make the library name relocatable
If a consumer links against a library, it references that library by the
id attribute in that library.  If the id attribute contains a absolute
path, then it will be used as the full path to link against breaking
relocatable builds entirely.

Fix by always making the id of the generated library be of the form
@rpath/libname.dylib.  This also requires that users of the library set
up an appropriate rpath entry.

Fixes https://gitlab.freedesktop.org/gstreamer/cerbero/-/issues/362

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/807>