cerbero-uninstalled: Allow Arch Linux to use ./cerbero-uninstalled
In Arch Linux the path `/usr/bin/pacman` exists and this led the script to
think it's been running in a Windows without the appropriate MSYS
environment. Use the bash OSTYPE env var to ensure we're in a msys OS
before do the MSYS checks.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/847>