cerbero: Redefine prefix value in pkgconfig files
People that use the official binaries on macOS and Windows keep
reporting issues saying that the prefix value in the pkgconfig file
points to a non-existent path, pointing to the build prefix on the
machine that the .pkg or .msi was built on.

Change all those to explicitly be a static non-existent path that also
contains a message telling users what to do if they see it in their
compiler command lines.

This is not needed on Linux because we do not ship official binaries
for that, and also because it might break someone's cross-compile
workflow since people build packages themselves on Linux.

Fixes https://gitlab.freedesktop.org/gstreamer/cerbero/-/issues/371

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/850>