libpng: fix failure to find correct zlib headers
Workaround the libpng build system mangling include paths by adding the
include path to the sysroot include directory by hand. In theory this
should not be needed, but seems to fix two issues: finding zlib.h when
there is no other than the one from the sysroot, and picking the right
zlib.h header (from the sysroot) when there is also a header in the host
(which can result in mismatching headers being used, and libpng really,
really does not like that).

There are chances that this fix is needed also for platforms other than
Android, but it's only tested for Android targets for now.