cerbero: Allow fetching bootstrap files without a checksum
Setting tarball_checksum = False now disables checksum verification
for bootstrap files.

This is *only* for very specific use-cases. One such use-case is the
Rust bootstrapper (added in this commit series), which involves
downloading the Rust package manifest toml, parsing it to get package
URLs and checksums, and adding those to the list of files to download.

However, when rustup actually uses these files, it will also download
.sha256 and .asc files for verification, and these are not listed in
the channel file. It is completely useless to check the checksum of
these, and trying to do so would also completely break the workflow.

So, we allow non-verification of checksums in this very special case.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/915>