cerbero: Implement offline support for Cargo recipes
To do this, we need to run cargo update to download everything into
the Cargo fetch cache and generate a Cargo.lock. Then we run cargo
vendor which will read Cargo.lock and vendor everything that's needed
into the sources directory.

This means we need to also extract the recipe during the fetch phase,
which can be confusing but them's the breaks.

For the sake of keeping the architecture simple, we allow cerbero to
extract the recipe twice (and run cargo vendor twice, which will be
a no-op).

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/915>