cerbero: Add support for cargo-c recipes
Used for building projects that use Cargo-C, such as gst-plugins-rs.

Cargo-C is installed as a build-tools recipe of type BuildType.CARGO,
which is enforced to only be usable by build-tools recipes.

We also fetch cargo-c as part of bootstrap because we need to run
cargo vendor to actually fetch everything needed. This required some
significant conversion of fetch functions to async.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/915>