cerbero/cargo: Don't require the system to provide a toml module
Now we download, extract, and use the tomli python module if the
system doesn't have any of: toml, tomli, tomlkit python modules.

We require only Python 3 and a compiler toolchain on Linux and macOS,
and we should keep it that way.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/915>