Split debuginfo generated by rustc to reduce plugin sizes
Without this, each cdylib plugin is hundreds of MBs. We can only do
this on macOS and MSVC. For other platforms, this option is unstable
and causes rustc to crash if you force-enable it. So we have to strip
the debuginfo on those platforms.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/915>