Download using powershell on Windows and rework download func
cerbero gen-cache was hanging on the MSVC post-merge CI intermittently
for some time, and recently it has started hanging consistently.
Debugging on the runner revealed that urllib2 was stuck somewhere that
could not be determined while downloading the artifact deps log, and
that it wasn't reading or writing anything. Most recently:

https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/28839005

urllib2 has always been slow and buggy, so let's just get rid of it
entirely in favour of Invoke-WebRequest in powershell.

The only side-effect of this is that we now require curl or wget on
non-Windows platforms, which is not a problem.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/943>