cerbero: Check that Cargo.toml version matches recipe version
Rust recipes that track git can easily go out of sync with the source
version, so add a check in the configure step for it.

For this, we need the toml python module, so we need to refactor the
code that finds it. It needs to move from the rust bootstrapper to
cerbero's config object. It also then needs to be in a fixed directory
inside the rust prefix, which means you need to re-bootstrap to get
this working.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/955>