gst-plugins-rs: version mismatch is error only when tagged_for_release
Checking for origin/main is not enough, because once we start making
development releases, gst-plugins-rs starts tracking a stable branch
of gst-plugins-rs as soon as it becomes available. A version change in
there means that we start failing CI for no reason.

We only need to update the version when tagging for a release, at
which time we will affix to a tag. In order to check that condition,
we need to move the check to the recipe.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1016>