lame: build with meson
Based on wrapdb wrap 3.100-7 with fixes for exported symbols.

Might not build asm for i386, but that's probably not too bad
seeing that it's a legacy platform and no assembly is shipped
for other architectures either.

Also removes the 'lame' binary from the package. We don't need
to ship that.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1111>