cerbero: Recipes can now download meson subprojects
The purpose is entirely for subprojects that are used only at build
time, either as 'copylibs' that are built as part of the project or
built and statically linked into the project.

pkg-config will be using this to download and build glib. It cannot
pick that up from the cerbero prefix because... that requires
pkg-config.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1173>