cerbero: Don't extract if already extracted in fetch
Some recipes need to extract the source tree in order to know what
else needs to be fetched. At present, this is true for Cargo recipes
where we need to run `cargo vendor` to fetch sources for offline
build, and for Meson recipes that require subprojects with wrap files,
where we need to download sources by reading the wrap files.

This is especially important for Cargo recipes, because `cargo vendor`
can take a dozen minutes or more on Windows. This should speed up the
MSVC CI by 15-20 minutes.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1192>