cargo: Restrict parallelism if a small system is detected
Cargo starves the rest of the system of RAM and wall clock cycles on
Intel Macs and other systems with low amount of RAM. So allow it to
use multiple jobs only if a big enough system is detected, and only
to the maximum number of threads.

For the purposes of Cerbero, more than 4 threads AND more than 8GB of
RAM is considered "big enough". The latter is only tested for macOS
through sysctl.

Fixes #446

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1293>