gst-plugins-rs: Super-strip plugins on iOS
Darwin based platform are the most strange I've ever seen. Due to their
mix of "flat" symbol exports (like ELF) and strong link-time symbol
resolution (like Windows), ld64 does not allow stripping symbols except
at link-time. Attempting to do so will either result in `strip`'s call
to `ld` complaining about undefined symbols, or a completely stubbed
library.

The obvious approach for this would be with -exported_symbols_list, but
there's a catch: Rust does not use the linker at all for this [1].
(Neither do CMake or Meson, fwiw.) This is a good approach for
staticlibs, BUT Apple provides another option which allows for sliding
a module definition file in: Single-Object Prelink [2] [3].

In Xcode, what this option does is:

- link world + dog into a new object file (`ld -r`) with all the flags
  necessary, just like a dylib
- then pass the file over to `libtool -static` (with extra,
  Xcode-specific information)

This commit replicates the above by:

- parsing the output of `nm` to figure out which GStreamer plugin APIs
  are exported, and storing these into a definition file
- unpacking the archive file into a temporary directory
- relinking with `-exported_symbol_lists` + definition file
- recreating a newly stripped static plugin with `libtool`.

This commit introduces a 10:1 saving on the biggest plugin of all,
libgstrswebrtc.a:

- debug: 504 MB -> 93 MB
- release: 144 MB -> 18 MB

Do notice that I don't strip any debugging information in the relinking.

[1]: https://github.com/rust-lang/rust/blob/58b4ebcdd62d82789ac50378698280006178d5fd/compiler/rustc_codegen_ssa/src/back/link.rs#L506
[2]: https://developer.apple.com/documentation/xcode/build-settings-reference
[3]: https://orangejuiceliberationfront.com/hiding-symbols-in-static-libraries-with-xcode-or-cmake/

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1255>