gst-plugins-rs: Enable superstripping on macOS too
This hadn't actually been tested on the CI before, which revealed a ton
of semantic issues between raw `subprocess.run` or
`subprocess.check_call` calls and the equivalent `cerbero.shell`
functionality. The main issue was that not only was the glob never
parsed, the output was being happily ignored.

Co-authored-by: Nirbheek Chauhan <nirbheek@centricular.com>
Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1341>