shell/new_call: Fix output streams contents being hidden from the user on error
subprocess.check_call, for some reason, does not log appropriately the contents
of stdout on error, nor it crashes out when a shell is required.

This is the cause of a hidden failure when issuing `ld` to prelink all objects
on macOS, because I specified a glob that would be expanded by the shell, and
this step never occurs because it's already passed verbatim to `sh`.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1364>