cerbero, config/android, openssl: Don't pass the Android toolchain path wholesale
In !1191 (commit 643087f3f09fd637b29335efb608ac7a82031ecf), the removal of GCC
support relied on making the Android compiler available to all processes. This
fixed accessing the host's compiler for Meson based builds like fribidi's, and
also allowed OpenSSL to build, but there was a hidden side effect: it broke
Rust's ability to compile build scripts because the linker it finds (LLVM 14
from the NDK) no longer supports `--no-add-needed` [1] [2].

A more Meson-ese fix is to pass the toolchain as a constant in the native file,
and fill the path in manually. This also needs to be injected manually into the
`PATH` for OpenSSL as that's how its configure script consumes the toolchain.

[1]: https://github.com/llvm/llvm-project/issues/54756
[2]: https://github.com/llvm/llvm-project/commit/815a1207bfe121c8dcf3804a4f4638e580f63519

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1368>