Fix soundtouch recipe include path
When built with autotools or cmake, the libsoundtouch project defines
the path '-I${includedir}/soundtouch' in its .pc file. This is the same path
that is also used in Ubuntu and Fedora Linux distributions packages.
This patch adds this path to the meson generated project using the
version 2.3.2-4 of the meson wrap for libsoundtouch.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1463>