WiX: implement Wine compatibility through an on-the-fly conversion wrapper
Ninja works in the native userspace, so it must have access to the POSIX
paths. But Windows style paths are needed for WiX, so I use special
flags to hoodwink the command line and tell the wrapper which arguments
need conversion.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1467>