recipe: Work around MetaUniversalRecipe shims not covering Darwin steps
BaseUniversalRecipe's __getattr__ yields over to the proxy recipe for
any nonexisting methods. This is no problem for the merge step, as
it's defined in the universal recipe directly, but steps such as
codesign and relocation are instead executed only in whatever
recipe was designated as the proxy.

Due to how recipes are loaded (I think it follows filesystem's sort
order), this recipe will usually be arm64 -- x86_64 libraries and
executables being left out of the relocation, and thus resulting
in a half-baked universal binary.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1499>