libvpx.recipe: Also build a shared lib on macOS
We want both dynamic and static libraries on macOS, and only static
libs on iOS, Android, Windows.

We used to have a static library on all platforms because libvpx's
custom build system was too finicky regarding shared builds. With the
port to Meson, for some reason the libvpx recipe was specifying *only*
dylibs on macOS, which is not intended.

We continue to require static libs on Windows because libvpx only
supports those.

Co-Authored-by: L. E. Segovia <amy@centricular.com>
Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1517>