ci, WiX: update packaging and recipe to v5
Luckily for us, our use of WiX is covered enough by Wine's existing
support. It only needs a bit of help with the compression level, which
has been added in this commit.

This also fixes the deployment of Wine-based MSIs which needs the
Wine environment variables properly set to the active prefix.

This also updates the image tag so that WiX v5 is correctly invoked.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1555>