WiX: port Visual Studio templates to v5 and fix their deployment
The files are actually deployed in (SDK)/share/vs/2010 as part of
both vsintegration-1.0 and the individual template features. The
/InstallTemplates action for the Visual Studio installation was
not called either.

This commit fixes both by ensuring the files are correctly
deployed into the present installations of Visual Studio, and
removing the duplicated files from the integration feature.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1555>