cerbero: Switch from wget to curl on Fedora 40 and newer
Fedora 40 switched from wget to wget2, and doesn't backport critical
bugfixes such as returning a non-zero status code on error, which can
happen due to any number of reasons, such as certificate
incompatibilities, network issues, domain not existing, server issues,
etc.

In these cases we get a success code and we don't try the fallback URL
and we fail in extract() with a confusing error message saying that
the downloaded tarball was not found.

Switch to curl, which doesn't have such issues.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1590>