osxrelocator: Fix RPATHs from libexec executables
There's an error being carried over from before !1485: the depth
calculation needs to be done folder-to-folder, so it was always off by
one. This error had been papered over by preserving the /lib/ folder in
the dylib ID, but since !1485 it was revealed by breaking running the
GStreamer executables in libexec.

This commit fixes it by properly doing the depth calculation. I've also
renamed the `lib_prefix` parameter as it is in fact the install prefix.

Fixes #499

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1610>