librsvg: Work around Apple ld64 bug with DWARF compact unwind info
ld -r strips rust_eh_personality because it doesn't realise that the
function is used by the DWARF unwind information. LLD 15+ does not have
the issue, but in turn is not able to do prelinking because that's an
Apple-proprietary feature.

See https://alanwu.space/post/symbol-hygiene/#the-return-of-rust_eh_personality

Fixes #507

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1670>