android: Don't ignore text relocation errors on 32-bit
We actually need to error out when we encounter text relocations when
creating a shared library from static libraries, since Android will
reject them and panic when the app is run.

`-Wl,-z,notext` will add a TEXTREL flag to the ELF even if there are
no object files with text relocations linked in, so it should
absolutely not be used!

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1845>