osx: Add packaging support for an iOS xcframework artifact
xcodebuild doesn't work on static frameworks, it needs a static
library (that can be a symlink, unlike the Headers folder).

The tarballs can be later used for melding the output for iOS and iOS
Simulator into a full xcframework in the next commit.

Support for other targets like tvOS and visionOS can also be added in
the future.

Co-authored-by: Nirbheek Chauhan <nirbheek@centricular.com>
Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1900>