cmake: Do not rely on the CERBERO_PREFIX environment variable
Use -DCMAKE_FIND_ROOT_PATH=%(prefix)s to pick the value for the sysroot
prefix, instead of relying on the $CERBERO_PREFIX environment variable.
This always works as expected, even if there is no intermediate shell
that would perform the expansion of the environment variable.

The following can be used to verify that CMake will not read the
environment variable when $CERBERO_PREFIX is the literal value, when
shell expansion is not involved:

    % cat > CMakeLists.txt <<EOF
    cmake_minimum_required (VERSION 3.18)
    set(ENV_VALUE "$ENV{VALUE}")
    message("VALUE_FROM_ENV: '${VALUE_FROM_ENV}'")
    message("ENV{VALUE}    : '${ENV_VALUE}'")
    EOF
    % VALUE=42 cmake -B build '-DVALUE_FROM_ENV=$VALUE'
    ...
    VALUE_FROM_ENV: '$VALUE'
    ENV{VALUE}    : '42'
    ...
    %

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1887>