Replace use_system_libs with SystemRecipes
SystemRecipes are special recipes that copy the .pc file into the
prefix so that the rest of the recipes can use it.

This system is safer than use_system_libs because it does not allow
access to all system libraries, only the necessary ones.

The commit adds three new properties:

* system_recipes_dir: system-recipes by default. It can not be inside recipes to import separately.
* allow_system_recipes: disable system recipes.
* system_recipes_priority: 1 to override normal recipes, 0 to be used as fallback.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1971>