cerbero: Try msys bsdtar, msys gnutar, then tarfile
On Windows, the tarfile Python module's implementation is very buggy,
and if Cerbero is invoked on a drive that is not C:\, extracting fails
because the source file and dest directory will be on different
drives, which `os.path.commonpath()` doesn't like.

But there's no winning, because it *also* happens if
`os.path.commonpath()` thinks the drives are different when they are
not, for example if one of the paths is an NT long path and the other
is not.

The only reason why we use tarfile is for some tarballs with symlinks,
because the various tar commands can't handle some types of symlinks
correctly in some cases. There are three known cases:

1. Relative path symlinks. These only work when the user has
   permissions to create symlinks.
2. Relative path symlinks can fail even when the user has permissions,
   such as with pixman.
3. Relative path symlinks that point to non-existent files, such as
   with soundtouch.
4. Absolute path symlinks: these always cause a failed extract, but
   we don't see these in tarballs.

These cases are supported at varying levels between win32 tar (which
is bsdtar), msys2's bsdtar, msys2's gnu tar. win32 tar hangs when
decompressing xz, so we try the other two.

(cherry picked from commit b2c23af96d1d105344b86f54e33588e3dcf72968)

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1984>