hacks: Fix breakage on Python 3.14
When building on a drive other than C:\, pathlib.Path.glob() was
returning zero results due to our overriding os.path.join. Un-override
it for the pathlib module.

This is building hacks on top of hacks, ideally we should get rid of
the os.path.join overriding hack and switch to pathlib for path
joining in Cerbero.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1993>