recipe: Add SystemRecipe.min_version and validate at creation
We need to validate because when system_recipes_priority is 1, we need
to use normal recipes as a fallback when the system package is not
installed. The error with the system recipes must be produced in the
constructor or in prepare(), not in the install step.

Example:
```
class Recipe(recipe.SystemRecipe):
    name = 'openssl'
    min_version = '3.0'
    pkgconfig_files = ['libcrypto', 'libssl', 'openssl']
```

System recipes will not be used when cross-compiling.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1971>